move extraction functions scripts > src
