+++ src/worldcereal/extract/patch_worldcereal.py
@@ -1,11 +1,12 @@
 """Extract WorldCereal preprocessed inputs using OpenEO-GFMAP package."""
 
+import copy
 import json
 import shutil
 from datetime import datetime
 from importlib.metadata import version
 from pathlib import Path
-from typing import Dict, List, Tuple, Union
+from typing import Dict, List, Optional, Tuple, Union
 
 import geojson
 import geopandas as gpd
@@ -32,14 +33,23 @@ from worldcereal.openeo.preprocessing import (
 )
 from worldcereal.utils.geoloader import load_reproject
 
-from worldcereal.extract.common import (  # isort: skip
+from worldcereal.extract.utils import (  # isort: skip
     get_job_nb_polygons,  # isort: skip
     pipeline_log,  # isort: skip
     upload_geoparquet_artifactory,  # isort: skip
+    S2_GRID,  # isort: skip
 )
 
 
 WORLDCEREAL_BEGIN_DATE = datetime(2017, 1, 1)
+
+
+DEFAULT_JOB_OPTIONS_PATCH_WORLDCEREAL = {
+    "executor-memory": "1800m",
+    "python-memory": "3000m",
+    "soft-errors": "true",
+    "max_executors": 22,
+}
 
 
 def create_job_dataframe_patch_worldcereal(
@@ -124,9 +134,7 @@ def create_job_patch_worldcereal(
     connection: openeo.DataCube,
     provider,
     connection_provider,
-    executor_memory: str,
-    python_memory: str,
-    max_executors: int,
+    custom_job_options: Optional[Dict[str, Union[str, int]]] = None,
 ) -> openeo.BatchJob:
     """Creates an OpenEO BatchJob from the given row information."""
 
@@ -189,12 +197,11 @@ def create_job_patch_worldcereal(
     number_polygons = get_job_nb_polygons(row)
     pipeline_log.debug("Number of polygons to extract %s", number_polygons)
 
-    job_options = {
-        "executor-memory": executor_memory,
-        "python-memory": python_memory,
-        "soft-errors": "true",
-        "max_executors": max_executors,
-    }
+    # Set job options
+    job_options = copy.deepcopy(DEFAULT_JOB_OPTIONS_PATCH_WORLDCEREAL)
+    if custom_job_options:
+        job_options.update(custom_job_options)
+
     return cube.create_job(
         out_format="NetCDF",
         title=f"GFMAP_Extraction_WORLDCEREAL_{s2_tile}_{valid_time}",
@@ -327,9 +334,6 @@ def post_job_action_patch_worldcereal(
     job_items: List[pystac.Item],
     row: pd.Series,
     extract_value: int,
-    description: str,
-    title: str,
-    spatial_resolution: str,
 ) -> list:
     """From the job items, extract the metadata and save it in a netcdf file."""
     base_gpd = gpd.GeoDataFrame.from_features(json.loads(row.geometry)).set_crs(
@@ -373,10 +377,10 @@ def post_job_action_patch_worldcereal(
             "valid_time": valid_time,
             "GFMAP_version": version("openeo_gfmap"),
             "creation_date": datetime.now().strftime("%Y-%m-%d %H:%M:%S"),
-            "description": description,
-            "title": title,
+            "description": "WorldCereal preprocessed inputs",
+            "title": "WorldCereal inputs",
             "sample_id": sample_id,
-            "spatial_resolution": spatial_resolution,
+            "spatial_resolution": "10m",
             "s2_tile": s2_tile,
             "_FillValue": 65535,  # No data value for uint16
         }
@@ -392,7 +396,6 @@ def generate_output_path_patch_worldcereal(
     job_index: int,
     row: pd.Series,
     asset_id: str,
-    s2_grid: gpd.GeoDataFrame,
 ):
     """Generate the output path for the extracted data, from a base path and
     the row information.
@@ -400,7 +403,7 @@ def generate_output_path_patch_worldcereal(
     sample_id = asset_id.replace(".nc", "").replace("openEO_", "")
 
     s2_tile_id = row.s2_tile
-    epsg = s2_grid[s2_grid.tile == s2_tile_id].iloc[0].epsg
+    epsg = S2_GRID[S2_GRID.tile == s2_tile_id].iloc[0].epsg
 
     return (
         root_folder
